requirements.txt

Flask==3.0.0
psycopg2==2.9.9
python-dotenv==1.0.0